final refactor , post MVP feature blocking, unit refactor renaming
