+++ python/framework/workers/abstract_blackbox_worker.py
