+++ python/framework/workers/core/heavy_workers.py
@@ -12,7 +12,7 @@ from typing import Dict, List
 import numpy as np
 
 from python.framework.types import Bar, TickData, WorkerContract, WorkerResult, WorkerType
-from python.framework.workers.abstract.abstract_blackbox_worker import \
+from python.framework.workers.abstract_blackbox_worker import \
     AbstractBlackboxWorker
