+++ python/framework/workers/core/rsi_worker.py
@@ -3,7 +3,7 @@ from typing import Dict, List
 import numpy as np
 
 from python.framework.types import Bar, TickData, WorkerContract, WorkerResult, WorkerType
-from python.framework.workers.abstract.abstract_blackbox_worker import \
+from python.framework.workers.abstract_blackbox_worker import \
     AbstractBlackboxWorker
